Mst: Added escape() and unescape() functions.
These functions escape and unescape backslash, newline ("\n"), and
form feed ("\f").

MST messages use newline for newlines (obviously), and form feed to
indicate a pause where the user has to press a button to continue
reading the message.

These should be escaped when editing to make it easier for the user
to manipulate things.

saveXML(): Escape string text.