Mst::saveMST(): Deduplicate string names.
If a string name is the same as the string table name, it's deduplicated.

msg_mainmenu.e.mst can now be converted to XML and back to MST, and the
resulting MST is identical to the original.

TODO: Do we need to deduplicate all strings, or just the string table name?