Mst::saveXML(): Save the per-message differential offsets in each <message> element.
Also for the initial WTXT/message count values in <mst06>.

For msg_mainmenu.e.xml: It seems that "AA" appears for messages that have
a button placeholder ('$'), and "A" appears for the button image string
immediately after that message. I might be able to use this information
later to properly generate the differential offset table instead of
including it in the XML file.

TODO: Remove <diffOffTbl> and use the per-message values.