Mst: Use the offset table directly when loading; offsets  are in triplets.
The three offsets:
- String name offset (Shift-JIS)
- String text offset (UTF-16)
- Placeholder name offset (Shift-JIS, may be zero)

Placeholder name offset is usually 0. This is still present in the
offset table, but the differential offset table *skips* it by using
'B'. The differential offset table is useless to us, but it's required
in order for Sonic'06 to parse it correctly and not crash.

saveXML(): The placeholder string is now saved as an attribute in the
relevant message instead of as its own message.

TODO: Update loadXML() and saveMST().