Mst::saveMST(): Moved `u16string msg_text` outside of the loop.
This eliminates having to create and destroy the u16string every
iteration.

Added msg_text.clear() to the case where byteswapping is needed but the
string is empty to ensure that it gets cleared. We were previously
depending on the string being reinitialized on every iteration for the
string to be cleared.